new activity flow layout and filtering
This batch of Reactome models is intended for display and interaction within the 1.0 release of Noctua.  It focuses on enabling all elements of reactions/functions/activities to be folded into single nodes using the available methods in Noctua1.0.  These nodes are then layed out such that causal relationships between the activities can be seen in the editor.  In order to enable folding, attributes of the targets of folded relationships including has_location assertions on physical entities and has_part assertion on complexes have been removed.  Each model represents a single pathway (biological process) with each activity node being a 'part_of' that pathway.  The pathway has_part activity assertions have also been removed as they obscure the view and provide no information beyond what is obvious.  The pathway node is present at the top left of the view as a label.  If a pathway contains other pathways, these relationships are kept and can be seen 'folded' into the the pathway node.