update to processing of sets
Sets can now be identified as active units of complexes and treated accordingly.  Likewise sets buried inside complexes are now maintained as union entities and attached directly to the parent complex.  Previously, if a set was found within a complex it would be broken down into its protein parts and all of these would be added as parts of the parent complex.