new reactome model collection - entity classes from new ontology
In this conversion, each physical entity node is typed with a class from an ontology induced from the Reactome physical entity set.  New classes are created for proteins (as well as complexes, sets, small molecules, etc.).  Where there are known equivalencies to entities in neo (e.g. uniprot), owl:equivalence axioms are used.  (All axioms are in a separate ontology, not included with this commit).  That ontology can be updated independently of these models.