Updating from new Reactome release
This version is working with the Reactome entity introduced in the last release.  Its an update of the content from Reactome.  No modeling changes.  Depends on https://github.com/geneontology/pathways2GO/raw/master/exchange/generated/Reactome_physical_entities.owl  to exist and be referenced in go-lego.owl for Minerva to work with this data.